+++ apps/ledger-live-desktop/src/newArch/features/Collectibles/Ordinals/components/Inscriptions/DetailsDrawer/Actions.tsx
@@ -1,9 +1,17 @@
 import React from "react";
-import { Flex, Icons, IconsLegacy, Text } from "@ledgerhq/react-ui";
+import { Flex, Icons, Text } from "@ledgerhq/react-ui";
 import Button from "~/renderer/components/Button";
 import { useTranslation } from "react-i18next";
+import { ExternalViewerButton } from "LLD/features/Collectibles/components/DetailDrawer/components";
+import { SimpleHashNft } from "@ledgerhq/live-nft/api/types";
+import { Account } from "@ledgerhq/types-live";
 
-const Actions = () => {
+type ActionsProps = {
+  inscription: SimpleHashNft;
+  account: Account;
+};
+
+const Actions: React.FC<ActionsProps> = ({ inscription, account }) => {
   const { t } = useTranslation();
   return (
     <Flex alignItems="center" columnGap={12}>
@@ -23,9 +31,11 @@ const Actions = () => {
           </Text>
         </Flex>
       </Button>
-      <Button outlineGrey>
-        <IconsLegacy.OthersMedium />
-      </Button>
+      <ExternalViewerButton
+        nft={inscription}
+        account={account}
+        metadata={inscription.extra_metadata}
+      />
     </Flex>
   );
 };