+++ apps/ledger-live-desktop/src/newArch/features/WalletSync/hooks/useOnTrustchainRefreshNeeded.ts
@@ -1,8 +1,12 @@
 import { useCallback } from "react";
 import { useDispatch } from "react-redux";
-import { MemberCredentials, Trustchain, TrustchainSDK } from "@ledgerhq/trustchain/types";
-import { setTrustchain, resetTrustchainStore } from "@ledgerhq/trustchain/store";
-import { TrustchainEjected } from "@ledgerhq/trustchain/errors";
+import {
+  MemberCredentials,
+  Trustchain,
+  TrustchainSDK,
+} from "@ledgerhq/ledger-key-ring-protocol/types";
+import { setTrustchain, resetTrustchainStore } from "@ledgerhq/ledger-key-ring-protocol/store";
+import { TrustchainEjected } from "@ledgerhq/ledger-key-ring-protocol/errors";
 import { log } from "@ledgerhq/logs";
 
 export function useOnTrustchainRefreshNeeded(