+++ apps/ledger-live-desktop/tests/enum/Account.ts
@@ -19,6 +19,42 @@ export class Account {
     "bc1q7ezsfc44adw2gyzqjmwhuh2e83uk8u5hrw590r",
   );
 
+  static readonly BTC_LEGACY_1 = new Account(
+    Currency.BTC,
+    "Bitcoin Legacy 1",
+    "1FMx2XwRHKXhbJcaekgPqN111wtRftkHyw",
+  );
+
+  static readonly BTC_LEGACY_2 = new Account(
+    Currency.BTC,
+    "Bitcoin Legacy 2",
+    "1Nen8hLiZysV6TWi6o6hTBp6dtkRXrJyWh",
+  );
+
+  static readonly BTC_SEGWIT_1 = new Account(
+    Currency.BTC,
+    "Bitcoin Segwit 1",
+    "3C7fQ47BiZuZN7V2WTgHLq4sqpb5BEp91i",
+  );
+
+  static readonly BTC_SEGWIT_2 = new Account(
+    Currency.BTC,
+    "Bitcoin Segwit 2",
+    "3B5psxZfUkU6AzPJoirgPQS7dC9vv7ohcK",
+  );
+
+  static readonly BTC_TAPROOT_1 = new Account(
+    Currency.BTC,
+    "Bitcoin Taproot 1",
+    "bc1pv4aytu7u4pk4nvelymxp65vndxqh4e3xsn0v7cguy6t98k5vynks5td23j",
+  );
+
+  static readonly BTC_TAPROOT_2 = new Account(
+    Currency.BTC,
+    "Bitcoin Taproot 2",
+    "bc1pywrw64mkvpzxkje5fkcz7vafj03yyzyvx4vg8wy2klvq8wtyfe7sq2jgzd",
+  );
+
   static readonly tBTC_1 = new Account(
     Currency.tBTC,
     "Bitcoin Testnet 1",
@@ -53,6 +89,12 @@ export class Account {
     Currency.ETH,
     "Ethereum 2",
     "0x43047a5023D55a8658Fcb1c1Cea468311AdAA3Ad",
+  );
+
+  static readonly ETH_3 = new Account(
+    Currency.ETH,
+    "Ethereum 3",
+    "0x50F293e4C3815F5214dac1Bae8F8FC3F24b8BA59",
   );
 
   static readonly tETH_1 = new Account(
@@ -287,5 +329,13 @@ export class Account {
     Currency.POL_UNI,
     "Polygon 1",
     "0xB9051f83AC6e147924377BBEebd1Aa7aB43a67F6",
+  );
+
+  static readonly EMPTY = new Account(Currency.BTC, "Empty", " ");
+
+  static readonly ETH_2_LOWER_CASE = new Account(
+    Currency.ETH,
+    "Ethereum 2",
+    "0x43047a5023d55a8658fcb1c1cea468311adaa3ad",
   );
 }