+++ libs/coin-modules/coin-polkadot/src/types/model.ts
@@ -53,24 +53,7 @@ export type ExplorerExtrinsic = {
   };
 };
 
-/// cf. ExtrinsicPayloadValue
-export type CoreTransasctionInfo = {
-  // before runtime upgrade
-  address: string;
-  blockHash: string;
-  blockNumber: `0x${string}`;
-  era: `0x${string}`;
-  genesisHash: string;
-  method: `0x${string}`;
-  nonce: `0x${string}`;
-  signedExtensions: string[];
-  specVersion: `0x${string}`;
-  tip: `0x${string}`;
-  transactionVersion: `0x${string}`;
-  version: number;
-};
-
-export type TransasctionPayloadInfo = {
+export type TransactionPayloadInfo = {
   // after runtime upgrade
   address: string;
   blockHash: string;
@@ -87,5 +70,5 @@ export type TransasctionPayloadInfo = {
 
 export type CoreTransaction = {
   registry: TypeRegistry;
-  unsigned: CoreTransasctionInfo | TransasctionPayloadInfo;
+  unsigned: TransactionPayloadInfo;
 };