+++ libs/coin-modules/coin-solana/src/api/chain/validators/bignum.ts
@@ -1,7 +1,12 @@
-import { coerce, instance, string } from "superstruct";
+import { coerce, instance, number, string } from "superstruct";
 import { BigNumber } from "bignumber.js";
 
 export const BigNumFromString = coerce(instance(BigNumber), string(), value => {
   if (typeof value === "string") return new BigNumber(value, 10);
+  throw new Error("invalid big num");
+});
+
+export const BigNumFromNumber = coerce(instance(BigNumber), number(), value => {
+  if (typeof value === "number") return new BigNumber(value, 10);
   throw new Error("invalid big num");
 });