+++ libs/ledgerjs/packages/hw-app-near/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-near
 
+## 6.29.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.4
+
+## 6.29.4-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.4-next.0
+
 ## 6.29.3
 
 ### Patch Changes